Delete the channel Event Queue when the channel is left
When a user unSubscribe from a channel, the events queue is not
deleted. If the user subscribe again to the same channel with the same
event, the events will be fired twice.

The best way to fix this would probably to check if the event is
already registered in the queue before adding it. But again it the two
channel have the same name, but are completely different, this could be
helpful.